refactor(cron): resolve model preflight behind one policy call
prepareCronRunContext sat at the 700-line cap with a ~40-line inline loop
that walked preflight candidates while tracking three mutable locals
(selected candidate, its index, and the first unavailable result).

Move that into resolveCronPreflight alongside the candidate list it
consumes in run-fallback-policy, returning a closed ok/reason result so the
caller reads a decision instead of reconstructing one. The lazy
model-preflight loader moves to its only consumer.

The pre-run persistence failure policy stays in run-prepare: the function it
guards is also awaited from run-finalize and run-executor.
