+++ src/cron/isolated-agent/run-fallback-policy.ts
@@ -3,11 +3,17 @@ import { resolveModelCandidateChain } from "../../agents/model-fallback-candidat
 import type { ModelCandidate } from "../../agents/model-fallback.types.js";
 import { resolveAgentModelFallbackValues } from "../../config/model-input.js";
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
+import { createLazyImportLoader } from "../../shared/lazy-promise.js";
 import type { CronJob } from "../types.js";
 import {
   resolveEffectiveModelFallbacks,
   resolveSubagentModelFallbacksOverride,
 } from "./run-execution.runtime.js";
+import { logWarn } from "./run.runtime.js";
+
+const cronModelPreflightRuntimeLoader = createLazyImportLoader(
+  () => import("./model-preflight.runtime.js"),
+);
 
 /** Resolves cron model fallbacks, giving explicit payload fallbacks precedence over subagent/default policy. */
 export function resolveCronFallbacksOverride(params: {
@@ -74,4 +80,61 @@ export function resolveCronPreflightCandidates(params: {
     requestedRouteResolution: "resolved",
     fallbacksOverride,
   });
+}
+
+/** Selects the reachable candidate and retains only its remaining fallback chain. */
+export async function resolveCronPreflight(
+  params: Parameters<typeof resolveCronPreflightCandidates>[0],
+) {
+  const modelPreflightRuntime = await cronModelPreflightRuntimeLoader.load();
+  const preflightCandidates = resolveCronPreflightCandidates(params);
+  let { provider, model } = params;
+  let selectedPreflightCandidate: ModelCandidate | undefined;
+  let selectedPreflightCandidateIndex = -1;
+  let firstUnavailablePreflight:
+    | Awaited<ReturnType<typeof modelPreflightRuntime.preflightCronModelProvider>>
+    | undefined;
+  for (const [index, candidate] of preflightCandidates.entries()) {
+    const candidatePreflight = await modelPreflightRuntime.preflightCronModelProvider({
+      cfg: params.cfg,
+      provider: candidate.provider,
+      model: candidate.model,
+    });
+    if (candidatePreflight.status === "available") {
+      selectedPreflightCandidate = candidate;
+      selectedPreflightCandidateIndex = index;
+      break;
+    }
+    firstUnavailablePreflight ??= candidatePreflight;
+  }
+  if (!selectedPreflightCandidate && firstUnavailablePreflight?.status === "unavailable") {
+    return { ok: false as const, reason: firstUnavailablePreflight.reason };
+  }
+  const modelFallbacksOverride =
+    selectedPreflightCandidate &&
+    (selectedPreflightCandidate.provider !== provider || selectedPreflightCandidate.model !== model)
+      ? preflightCandidates
+          .slice(selectedPreflightCandidateIndex + 1)
+          .map((candidate) => `${candidate.provider}/${candidate.model}`)
+      : undefined;
+  // When preflight skips the first local candidate, start at the reachable provider.
+  if (selectedPreflightCandidate && modelFallbacksOverride) {
+    if (firstUnavailablePreflight?.status === "unavailable") {
+      logWarn(
+        `[cron:${params.job.id}] ${firstUnavailablePreflight.reason}; continuing with fallback ${selectedPreflightCandidate.provider}/${selectedPreflightCandidate.model}.`,
+      );
+    }
+    provider = selectedPreflightCandidate.provider;
+    model = selectedPreflightCandidate.model;
+  }
+  return {
+    ok: true as const,
+    provider,
+    model,
+    modelFallbacksOverride,
+    runtimePluginCandidates:
+      selectedPreflightCandidateIndex >= 0
+        ? preflightCandidates.slice(selectedPreflightCandidateIndex)
+        : preflightCandidates,
+  };
 }
