+++ src/cron/isolated-agent/run-prepare-runtime.ts
@@ -51,9 +51,6 @@ const cronExternalContentRuntimeLoader = createLazyImportLoader(
 const cronAuthProfileRuntimeLoader = createLazyImportLoader(
   () => import("./run-auth-profile.runtime.js"),
 );
-const cronModelPreflightRuntimeLoader = createLazyImportLoader(
-  () => import("./model-preflight.runtime.js"),
-);
 export async function loadSessionAccessorRuntime() {
   return await sessionAccessorRuntimeLoader.load();
 }
@@ -64,10 +61,6 @@ export async function loadCronExternalContentRuntime() {
 
 async function loadCronAuthProfileRuntime() {
   return await cronAuthProfileRuntimeLoader.load();
-}
-
-export async function loadCronModelPreflightRuntime() {
-  return await cronModelPreflightRuntimeLoader.load();
 }
 
 function hasConfiguredAuthProfiles(cfg: OpenClawConfig): boolean {
